scripts/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ieee-3335-tools"
version = "0.2.0"
description = "Tools for working with IEEE-3335 documents - extract and convert PDFs, DOCX, and PPTX files to markdown"
requires-python = ">=3.10"
dependencies = [
    "pypdfium2>=4.30.0",
    "rich>=14.2.0",
    "typer>=0.9.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.21",
    "marker-pdf>=0.2.0",
    "hatchling>=1.27.0",
    "pillow>=10.1.0,<11.0.0",
]
readme = "README.md"

[project.scripts]
p3335 = "ieee_3335_tools.cli:main"
ieee-3335-tools = "ieee_3335_tools.cli:main"
